pyproject.toml
[build-system]
# 40.8.0 is the first version compatible with PEP 517
requires = ["setuptools>=40.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "micromagneticmodel"
version = "0.61.0"
description = "Python-based domain-specific language for computational magnetism"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}

authors = [
    {name = "Marijan Beg"},
    {name = "Martin Lang"},
    {name = "Samuel Holt"},
    {name = "Hans Fangohr"}
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Visualization"
]

dependencies = [
    "discretisedfield~=0.61.0"
]

[project.optional-dependencies]
dev = [
    "build",
    "invoke",
    "nbval",
    "pre-commit",
    "pytest-cov",
    "twine"
]

[project.urls]
homepage = "https://ubermag.github.io"
documentation = "https://ubermag.github.io/documentation/micromagneticmodel"
repository = "https://github.com/ubermag/micromagneticmodel"



# to define additional information we use a custom table
# we need to own the entry for tool.NAME on PyPI
# see: https://www.python.org/dev/peps/pep-0518/#tool-table
[tool.ubermag]
doi = "10.5281/zenodo.3539479"
copyright_holder = "Marijan Beg and Hans Fangohr"
contributors = [
    "Thomas Kluyver",
    "Ryan A. Pepper",
    "Sergii Mamedov",
    "Natalie Downing"
]
about = """
- Domain-specific language for computational magnetism.
"""

[tool.coverage.run]
omit = ["micromagneticmodel/tests/*"]

[tool.isort]
profile = "black"
skip_gitignore = true  # ignores files listed in .gitignore

[tool.black]
experimental-string-processing = true
